Add new frame deletion methods, guard file read/write, rewrite concat/split, add test cases and refactor

+++ src/InMemoryFwobFile.cs
@@ -1,13 +1,12 @@
 ﻿using Mozo.Fwob.Exceptions;
-using Mozo.Fwob.Models;
+using Mozo.Fwob.Abstraction;
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
 
 namespace Mozo.Fwob;
 
-public class InMemoryFwobFile<TFrame, TKey> : AbstractFwobFile<TFrame, TKey>
+public sealed class InMemoryFwobFile<TFrame, TKey> : AbstractFwobFile<TFrame, TKey>
     where TFrame : class, new()
     where TKey : struct, IComparable<TKey>
 {
@@ -17,11 +16,14 @@ public class InMemoryFwobFile<TFrame, TKey> : AbstractFwobFile<TFrame, TKey>
         set
         {
             if (value == null)
-                throw new ArgumentNullException(nameof(Title));
+                throw new ArgumentNullException(nameof(value));
+
             if (value.Length == 0)
-                throw new ArgumentException("Argument can not be empty", nameof(Title));
-            if (value.Length > FwobLimits.MaxTitleLength)
+                throw new ArgumentException("Argument can not be empty", nameof(value));
+
+            if (value.Length > Limits.MaxTitleLength)
                 throw new TitleTooLongException(value, value.Length);
+
             _title = value;
         }
     }
@@ -31,7 +33,6 @@ public class InMemoryFwobFile<TFrame, TKey> : AbstractFwobFile<TFrame, TKey>
     public InMemoryFwobFile(string title)
     {
         _title = title;
-        FrameInfo = FrameInfo.FromSystem(typeof(TFrame), typeof(TKey));
     }
 
     #region Implementations of IFrameQueryable
@@ -40,46 +41,78 @@ public class InMemoryFwobFile<TFrame, TKey> : AbstractFwobFile<TFrame, TKey>
 
     private readonly List<TFrame> _frames = new();
 
-    public override TFrame? GetFrame(long index)
+    protected override TKey InternalGetKeyAt(long index)
     {
-        if (index < 0 || index >= FrameCount)
-            return null;
+        return GetKey(_frames[(int)index]);
+    }
+
+    protected override TFrame InternalGetFrameAt(long index)
+    {
         return _frames[(int)index];
     }
 
-    private int GetBound(TKey key, bool lower)
+    public override IEnumerable<TFrame> GetFrames(IEnumerable<TKey> keys)
     {
-        int lo, hi;
-        for (lo = 0, hi = _frames.Count; lo < hi;)
+        ValidateKeys(keys, nameof(keys));
+
+        if (_frames.Count == 0)
+            yield break;
+
+        long idx = 0;
+        foreach (TKey key in keys)
         {
-            int mid = lo + (hi - lo >> 1);
-            int cmp = GetKey(_frames[mid]).CompareTo(key);
-            if (lower ? cmp < 0 : cmp <= 0)
-                lo = mid + 1;
-            else
-                hi = mid;
+            (long lb, long ub) = GetEqualRange(key, idx, _frames.Count);
+
+            if (lb == ub)
+                continue;
+
+            for (long i = lb; i < ub; i++)
+                yield return InternalGetFrameAt(i);
+
+            idx = ub;
         }
-        return lo;
     }
 
-    public override IEnumerable<TFrame> GetFrames(TKey firstKey, TKey lastKey)
+    public override IEnumerable<TFrame> GetFramesBetween(TKey firstKey, TKey lastKey)
     {
-        Debug.Assert(firstKey.CompareTo(lastKey) <= 0);
+        if (firstKey.CompareTo(lastKey) > 0)
+            throw new ArgumentException($"{nameof(lastKey)} must be greater than or equal to {nameof(firstKey)}");
 
-        for (int i = GetBound(firstKey, true); i < _frames.Count && GetKey(_frames[i]).CompareTo(lastKey) <= 0; i++)
+        if (_frames.Count == 0)
+            yield break;
+
+        long lb = GetLowerBound(firstKey, 0, _frames.Count);
+        long ub = GetUpperBound(lastKey, lb, _frames.Count);
+
+        while (lb < ub)
+            yield return _frames[(int)lb++];
+    }
+
+    public override IEnumerable<TFrame> GetFramesBefore(TKey lastKey)
+    {
+        if (_frames.Count == 0)
+            yield break;
+
+        long ub = GetUpperBound(lastKey, 0, _frames.Count);
+
+        for (int i = 0; i < ub; i++)
             yield return _frames[i];
     }
 
     public override IEnumerable<TFrame> GetFramesAfter(TKey firstKey)
     {
-        for (int i = GetBound(firstKey, true); i < _frames.Count; i++)
+        if (_frames.Count == 0)
+            yield break;
+
+        long lb = GetLowerBound(firstKey, 0, _frames.Count);
+
+        for (int i = (int)lb; i < _frames.Count; i++)
             yield return _frames[i];
     }
 
-    public override IEnumerable<TFrame> GetFramesBefore(TKey lastKey)
+    public override IEnumerable<TFrame> GetAllFrames()
     {
-        for (int i = 0; i < _frames.Count && GetKey(_frames[i]).CompareTo(lastKey) <= 0; i++)
-            yield return _frames[i];
+        return _frames;
     }
 
     public override long AppendFrames(IEnumerable<TFrame> frames)
@@ -89,10 +122,12 @@ public class InMemoryFwobFile<TFrame, TKey> : AbstractFwobFile<TFrame, TKey>
 
         TFrame? last = _frames.LastOrDefault();
         long count = 0;
+
         foreach (TFrame frame in frames)
         {
             if (last != null && GetKey(frame).CompareTo(GetKey(last)) < 0)
                 throw new KeyOrderViolationException();
+
             _frames.Add(frame);
             last = frame;
             count++;
@@ -120,23 +155,93 @@ public class InMemoryFwobFile<TFrame, TKey> : AbstractFwobFile<TFrame, TKey>
         return count;
     }
 
-    public override long DeleteFramesAfter(TKey firstKey)
+    public override long DeleteFrames(IEnumerable<TKey> keys)
     {
-        int lb = GetBound(firstKey, true), len = _frames.Count - lb;
-        _frames.RemoveRange(lb, len);
-        return len;
+        ValidateKeys(keys, nameof(keys));
+
+        if (_frames.Count == 0)
+            return 0;
+
+        IEnumerator<TKey> it = keys.GetEnumerator();
+
+        long readerIdx = 0, writerIdx = 0;
+
+        while (readerIdx == writerIdx && it.MoveNext())
+        {
+            (writerIdx, readerIdx) = GetEqualRange(it.Current, readerIdx, _frames.Count);
+        }
+
+        if (readerIdx == writerIdx)
+            return 0;
+
+        while (it.MoveNext())
+        {
+            (long lb, long ub) = GetEqualRange(it.Current, readerIdx, _frames.Count);
+
+            while (readerIdx < lb)
+                _frames[(int)(writerIdx++)] = _frames[(int)readerIdx++];
+
+            readerIdx = ub;
+        }
+
+        int len = _frames.Count;
+
+        while (readerIdx < len)
+            _frames[(int)writerIdx++] = _frames[(int)readerIdx++];
+
+        _frames.RemoveRange((int)writerIdx, len - (int)writerIdx);
+
+        return len - _frames.Count;
+    }
+
+    public override long DeleteFramesBetween(TKey firstKey, TKey lastKey)
+    {
+        if (firstKey.CompareTo(lastKey) > 0)
+            throw new ArgumentException($"{nameof(lastKey)} must be greater than or equal to {nameof(firstKey)}");
+
+        if (_frames.Count == 0)
+            return 0;
+
+        int lb = (int)GetLowerBound(firstKey, 0, _frames.Count);
+        int ub = (int)GetUpperBound(lastKey, lb, _frames.Count);
+
+        _frames.RemoveRange(lb, ub - lb);
+
+        return ub - lb;
     }
 
     public override long DeleteFramesBefore(TKey lastKey)
     {
-        int len = GetBound(lastKey, false);
+        if (_frames.Count == 0)
+            return 0;
+
+        int len = (int)GetUpperBound(lastKey, 0, _frames.Count);
+
         _frames.RemoveRange(0, len);
+
         return len;
     }
 
-    public override void ClearFrames()
+    public override long DeleteFramesAfter(TKey firstKey)
     {
+        if (_frames.Count == 0)
+            return 0;
+
+        int lb = (int)GetLowerBound(firstKey, 0, _frames.Count);
+        int len = _frames.Count - lb;
+
+        _frames.RemoveRange(lb, len);
+
+        return len;
+    }
+
+    public override long DeleteAllFrames()
+    {
+        int len = _frames.Count;
+
         _frames.Clear();
+
+        return len;
     }
 
     #endregion
@@ -162,9 +267,7 @@ public class InMemoryFwobFile<TFrame, TKey> : AbstractFwobFile<TFrame, TKey>
 
     public override int AppendString(string str)
     {
-        if (_dict.TryGetValue(str, out int key))
-            return key;
-        _dict[str] = key = _data.Count;
+        int key = _dict[str] = _data.Count;
         _data.Add(str);
         return key;
     }
@@ -174,10 +277,12 @@ public class InMemoryFwobFile<TFrame, TKey> : AbstractFwobFile<TFrame, TKey>
         return _dict.ContainsKey(str);
     }
 
-    public override void ClearStrings()
+    public override int DeleteAllStrings()
     {
+        int len = _data.Count;
         _data.Clear();
         _dict.Clear();
+        return len;
     }
 
     #endregion
